% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_documents}
\alias{download_documents}
\title{Download PDF documents belonging to lawsuits}
\usage{
download_documents(data, path, login = NULL, password = NULL,
  only_petitions = FALSE, progress = FALSE)
}
\arguments{
\item{data}{A character vector with one or more lawsuit IDs or
the tibble returned by [get_metadata()] (can be filtered)}

\item{path}{Path to directory where to save files}

\item{login}{ESAJ system login (if left `NULL` and your're
not logged in yet, will ask for it)}

\item{password}{ESAJ system password (if left `NULL` and your're
not logged in yet, will ask for it)}

\item{only_petitions}{Whether to only download petitions}

\item{progress}{Whether to show a progress bar}
}
\description{
This function accesses lawsuits' ESAJ pages and
tries to download documents (if IDs are provided to `data`
and a lawsuit isn't found, shows error but skips it).
}
\seealso{
[get_metadata()]
}
