% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from documents belonging to lawsuits}
\usage{
get_metadata(id, login = NULL, password = NULL, only_petitions = FALSE)
}
\arguments{
\item{id}{A character vector with one or more lawsuit IDs}

\item{login}{ESAJ system login (if left `NULL` and your're
not logged in yet, will ask for it)}

\item{password}{ESAJ system password (if left `NULL` and your're
not logged in yet, will ask for it)}

\item{only_petitions}{Whether to only get petitions}
}
\description{
This function accesses lawsuits' ESAJ pages and
tries to get the metadata of all documents (if lawsuit
isn't found, shows error but skips it).
}
\seealso{
[download_documents()]
}
