% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{login_esaj}
\alias{login_esaj}
\title{Log into ESAJ system}
\usage{
login_esaj(login = NULL, password = NULL)
}
\arguments{
\item{login}{ESAJ system login (if left `NULL`, will ask for it)}

\item{password}{login ESAJ system login (if left `NULL`, will ask for it)}
}
\description{
Log into ESAJ system
}
