% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_aux.R
\name{visualize_page}
\alias{visualize_page}
\title{Vizualize web page from httr::GET}
\usage{
visualize_page(response)
}
\arguments{
\item{response}{Response from an HTTP request}
}
\description{
Vizualize web page from httr::GET
}
